% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotateBins}
\alias{annotateBins}
\title{annotateBins}
\usage{
annotateBins(bins, annotation, id.name = "id",
  column.name = "coding_genes", ignore.strand = TRUE)
}
\arguments{
\item{bins}{GRanges object with a disjoint partition of the genome}

\item{annotation}{GRanges object with annotations (e.g. genes, lncRNAs)}

\item{column.name}{Name of the field from annotation to be added as metadata column to the result GRanges.}

\item{ignore.strand}{Ignore strandedness for findOverlaps}
}
\value{
GRanges object with a copy of bins with an additional metadata column named "overlaps" with the annotations from annotation falling into each bin
}
\description{
This function loads two GRanges objects, one with a partition of the genome (disjoint GRanges object) and one annotation track with annotation
 of interest (e.g. genes, lncRNAs)
The function adds metadata to the bins GRanges object with the annotations falling in each bin of the partition.
}
\examples{
annotateBins(bins, annotation)
}
\keyword{GRanges}
\keyword{binning}
\keyword{hits}
