% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotateBinsFromGRanges}
\alias{annotateBinsFromGRanges}
\title{Annotate bins per category}
\usage{
annotateBinsFromGRanges(bins, annotations)
}
\arguments{
\item{bins}{GRanges object with a disjoint partition of the genome}

\item{annotations}{GRanges object with annotations (e.g. genes, lncRNAs)}
}
\value{
Annotation according to category
}
\description{
Annotate bins per category
}
