% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed2GRanges.R
\name{bed2GRanges}
\alias{bed2GRanges}
\title{Generate a GRanges object from a BED file}
\usage{
bed2GRanges(file, chr.field = "chr", start.field = "start",
  end.field = "end", strand.field = "strand")
}
\arguments{
\item{file}{Name of the BED file}

\item{chr.field}{Name in the header of the BED file associated with the chr field
for the GRanges conversion}

\item{start.field}{Name in the header of the BED file associated with the start field
for the GRanges conversion}

\item{end.field}{Name in the header of the BED file associated with the end field
for the GRanges conversion}

\item{strand.field}{Name in the header of the BED file associated with the strand field
for the GRanges conversion}
}
\value{
GRanges object
}
\description{
This function loads a BED-like file and stores it as a GRanges object.
The tab-delimited file must not be ordered, but must include a header line indicating the fields which includes chr.field, start.field,
end.field (minimum), and optionally: strand.field.
Any columns apart from the four columns described are loaded into the GRanges object as metadata.
}
\examples{
bed2Granges('my_bed_file.bed')
}
\keyword{BED}
\keyword{GRanges}
