% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStatistics.R
\name{computeStatistics}
\alias{computeStatistics}
\title{ComputeProbability}
\usage{
computeStatistics(gr, hits.column.name = "Lynch.wilcox.0.005.hits",
  CG.column.name = "CG.hits", N.column.name = "N.hits",
  N.rate.column.name = "N.prop", thrN = 0)
}
\arguments{
\item{gr}{GRanges object with a disjoint partition of the genome and
two metadata columns named as specified in hist.colum.name and CG.column.name}

\item{hits.column.name}{Number of hits in each bin}

\item{CG.column.name}{Number of dinucleotides CG per bin}

\item{N.column.name}{Number of N per bin}

\item{thrN}{Threshold for the fraction of N in each bin}

\item{N.prop.name}{Proportion of N in each bin}
}
\value{
Copy of GRanges object gr with new metadata columns "observed" "probability" and "expected" (previous metadata columns in gr are also kept)
}
\description{
Compute observed, expected and probability for a GRanges object given a counting for hits (features) and CGs.
Find bins with a number of signatuire significatly different from the expected,
calculated as N*p_0, where N is the total nuymber of bins, and the probability
p_0 is given by the number of CG in each bin divided by the total number
of CG in the reference genome
}
\examples{
computeStatistics (gr)
}
\keyword{GRanges}
\keyword{expected}
\keyword{observations}
\keyword{probability}
