% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{getGencodeGTF}
\alias{getGencodeGTF}
\title{Get Gencode GTF data from RAW GENCODE file}
\usage{
getGencodeGTF(file, genes.only = TRUE)
}
\arguments{
\item{file}{File with genome annotation}

\item{genes.only}{If TRUE (default) only genes are considered}
}
\value{
Annotation
}
\description{
Get Gencode GTF data from RAW GENCODE file
}
