% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges2bed.R
\name{GRanges2bed}
\alias{GRanges2bed}
\title{Generate a Granges object (only seqname, start and end) to a BED-like file (3 fields format)}
\usage{
GRanges2bed(gr, file, chr.header = "chr", start.header = "start",
  end.header = "end")
}
\arguments{
\item{gr}{GRanges object}

\item{file}{Name of BED file to be generated}

\item{chr.header}{Name of header field for chromosome}

\item{start.header}{Name of header field for start coordinate}

\item{end.header}{Name of Header field for end coordinate}
}
\value{
BED file
}
\description{
Generate a Granges object (only seqname, start and end) to a BED-like file (3 fields format)
}
\examples{
Granges2Bed(gr, 'my_bed_file.bed')
}
\keyword{BED}
\keyword{GRanges}
