% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitsPerBin.R
\name{hitsPerBin}
\alias{hitsPerBin}
\title{hitsPerBin}
\usage{
hitsPerBin(bins, features, column.name = "hits")
}
\arguments{
\item{bins}{GRanges object with a disjoint partition of the genome}

\item{features}{GRanges object with a list of loci of interest (e.g. genetic markers)}

\item{column.name}{Name of the metadata column to be added to the resulting GRanges with the number
of features falling into each bin}
}
\value{
GRanges object with a copy of bins including an additional metadata column 'hits'
 with the number features ranges falling into each bin.
}
\description{
This function loads two GRanges objects, one with a partition of the genome (disjoint GRanges object) and one with loci of interest (e.g. relevant biomarkers).
The function counts how many of the loci of interest (biomarkers) fall in each bin of the partition.
}
\examples{
hitsPerBin(subject, query)
}
\keyword{GRanges}
\keyword{binning}
\keyword{hits}
