% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGenome.R
\name{loadGenome}
\alias{loadGenome}
\title{loadGenome}
\usage{
loadGenome(genome.ver = "hg38", exclude.scaffold.chrs = TRUE)
}
\arguments{
\item{genome.ver}{Reference genome (hg18, hg19, hg38)}

\item{exclude.scaffold.chrs}{If true (default), do not include scaffold chromosomes in the returned DNAStringSet object}
}
\value{
DNAStringSet object with the chosen reference genome associated to genome.ver.
By default, the sequence of the '+' strand is downloaded.
}
\description{
Loads BSgenome sequences for working with genomic coordinates
}
\examples{
genome <- loadGenome (genome.ver = "hg38", exclude.scaffold.chrs = TRUE)

}
\keyword{Reference}
\keyword{genome}
