% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSignature.R
\name{loadSignature}
\alias{loadSignature}
\title{loadSignature}
\usage{
loadSignature(file, numberFeatures = 0, coord.origin = "hg38",
  coord.dest = "hg38")
}
\arguments{
\item{file}{Name of file containing signature coordinates (cvs format)}

\item{numberFeatures}{Limit the features to import from file to the first numberFeatures features;
if NULL (default) all the features in file are imported}

\item{coord.origin}{String with the reference genome version of the coordinates in file (supports values "hg38", "hg19", "hg18")}

\item{coord.dest}{If different than coord.origin, the coordinates in file will be converted from coord.origin to coord.dest.
Current support only for conversions from "hg38" to "hg19" and "hg18".}
}
\value{
GRanges object with the coordinates of the signature (extra columns are added as metadata to the GRanges object)
}
\description{
Loads a signature of genomic coordinates and convert to destination genome reference version (if required)
}
\examples{
signature <- loadSignature ()
}
\keyword{signatures}
