% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeHits.R
\name{mergeHits}
\alias{mergeHits}
\title{Gnerate GRanges object with number of signatures, number of dinucleotides CG
number of N and N proportion per bin}
\usage{
mergeHits(gr.hits, gr_CG.hits, gr_N.hits)
}
\arguments{
\item{gr.hits}{GRanges object with number of hits per bin as metadata}

\item{gr_CG.hits}{GRanges object with number of dinucleotides CG per bin as metadata}

\item{gr_N.hits}{Granges object with number of N per bin as metadata}
}
\value{
GRanges object
}
\description{
Gnerate GRanges object with number of signatures, number of dinucleotides CG
number of N and N proportion per bin
}
\examples{
mergeHits(gr.hits, gr_CG.hits, gr_N.hits)
}
\keyword{GRanges}
