% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.R
\name{resume}
\alias{resume}
\title{Put results in dataframe according to ascending p value}
\usage{
resume(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
data frame
}
\description{
Put results in dataframe according to ascending p value
}
\keyword{annotation}
\keyword{results,}
