% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceCoordinates.R
\name{sequenceCoordinates}
\alias{sequenceCoordinates}
\title{Searches for matches of a nucleotide sequence genome wide and returns matching coordinates}
\usage{
sequenceCoordinates(genome, pattern = "CG")
}
\arguments{
\item{genome}{DNAStringSet object with the reference genome (typically as provided by getSeq from package BSgenome)}

\item{pattern}{DNA sequence to search for matches.}
}
\value{
GRanges object with the coordinates of every match for the pattern sequence in the genome (genome version as provided in genome.ver)
}
\description{
Searches for matches of a nucleotide sequence genome wide and returns matching coordinates
}
\examples{
sequenceCoordinates("hg38", "CG")  #searches for dinucleotides CpG genome-wide in hg38.
}
\keyword{CG}
\keyword{GRanges}
