% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADbinning.R
\name{TADbinning}
\alias{TADbinning}
\title{TADbinning}
\usage{
TADbinning(file, genome, chr.field = "chr", start.field = "start",
  end.field = "end", strand.field = "strand",
  gap.consecutive.bins = 0, verbose = TRUE)
}
\arguments{
\item{file}{Name of BED file containing TAD coordinates}

\item{genome}{DNAStringSet object with the reference genome as returned by the function loadGenome}

\item{chr.field}{Field name in header associated to the chr field for the GRanges conversion}

\item{start.field}{Field name in header associated to the start field for the GRanges conversion}

\item{end.field}{Field name in header associated to the end field for the GRanges conversion}

\item{strand.field}{Field name in header associated to the strand field for the GRanges conversion}

\item{gap.consecutive.bins}{Size of the gap between two consecutive bins. I.e. if there are two bins a: [a1,a2], b: [b1,b2],
what is the value of b1-a2, so we consider the bins consecutive? (default 0). Set gap.consecutive.bins to 0 when consecutive gaps
share end/start coordinate (i.e. b1==a2))}

\item{verbose}{TRUE (default) plots and shows statistics on distributions of bins width}
}
\value{
GRanges object with a whole genome partition/binning defined by the provided TADs coordinates
}
\description{
This function loads a BED-like file and stores it into a GRanges object. This BED-like file contains bin (e.g TAD) coordinates.
The function prints basic statistics on the width and number of bins the genome is divided into.
The tab-delimited file must not be ordered, but must include a header line indicating the fields which includes chr.field, start.field,
end.field (minimum), and optionally: strand.field.
Any columns apart from the four columns described are loaded into the GRanges object as metadata.
}
\examples{
TADbinning('HiCtool_allchrs_TADS.bed', genome, chr.field="chr", start.field="start", end.field="end", gap.consecutive.bins=0, verbose=F)
}
\keyword{BED}
\keyword{GRanges}
\keyword{TAD}
\keyword{binning}
